+++ cse30life.h
@@ -1,4 +1,4 @@
-//! CSE30-WI21
+//! CSE30-SP21
 //! cse30life functions
 //! functions to support life!
 
@@ -25,35 +25,35 @@ extern double getSecs();
  *  predefined shapes
  */
 //! create a glider
-extern void glider(unsigned char *buf, uint32_t cols, uint32_t v, uint32_t h);   
+extern void glider(unsigned char *buf, size_t cols, size_t v, size_t h);   
 
 //! create a toad
-extern void toad(unsigned char *buf, uint32_t cols, uint32_t v, uint32_t h); 
+extern void toad(unsigned char *buf, size_t cols, size_t v, size_t h); 
 
 //! create an acorn
-extern void acorn(unsigned char *buf, uint32_t cols, uint32_t v, uint32_t h); 
+extern void acorn(unsigned char *buf, size_t cols, size_t v, size_t h); 
 
 //! create an oscillator horiz orient
-extern void oscillator(unsigned char *buf, uint32_t cols, uint32_t v, uint32_t h);
+extern void oscillator(unsigned char *buf, size_t cols, size_t v, size_t h);
 
 //! create an oscillator vertical 
-extern void oscillatorV(unsigned char *buf, uint32_t cols, uint32_t v, uint32_t h); 
+extern void oscillatorV(unsigned char *buf, size_t cols, size_t v, size_t h); 
 
 /**
  * board plotting/printing
  */
 //! print the board to the screen as ascii
-extern void printAsciiBoard(unsigned char *buf, uint32_t rows, uint32_t cols, uint32_t gen);
+extern void printAsciiBoard(unsigned char *buf, size_t rows, size_t cols, size_t gen);
 
 //! create a new graphics window
-extern int createWindow(uint32_t scale, uint32_t rows, uint32_t cols);
+extern int createWindow(size_t scale, size_t rows, size_t cols);
 
 //! close the graphics window
 extern int closeWindow();  // close the window
 
 //! print the board to a window
-extern void plotBoard(unsigned char *buf, uint32_t rows, uint32_t cols);
+extern void plotBoard(unsigned char *buf, size_t rows, size_t cols);
 
 //! dump board to a file
-extern void dumpBoard(unsigned char *buf, uint32_t rows, uint32_t cols, const char *dumpFileName);
+extern void dumpBoard(unsigned char *buf, size_t rows, size_t cols, const char *dumpFileName);
