+++ sim.h
@@ -2,13 +2,13 @@
  * Assignment: life
  * Name: TODO
  * PID: TODO
- * Class: UCSD CSE30-WI21
+ * Class: UCSD CSE30-SP21
  *
  */
 #ifndef _SIM_H
 #define _SIM_H
 #include "board.h"
 
-extern void simLoop(boards_t *self, uint32_t steps); // simulation loop
+extern void simLoop(boards_t *self, size_t steps); // simulation loop
 
 #endif
